widowx_envs/requirements.txt
pillow
protobuf
funcsigs
future
imageio
imageio-ffmpeg<0.5.0
matplotlib
moviepy
opencv-python
numpy==1.23.3
pyquaternion
scikit-learn
scikit-image
scipy==1.10.1
six
requests
nvidia_smi
rospkg
modern_robotics==1.1.1
gym
tqdm
transformations
ipdb
joblib
pickle5
h5py
funcsigs
numba
netifaces
pyspacemouse
git+https://github.com/rail-berkeley/oculus_reader.git
